requirements.txt
#Server
Django==1.4.3
gunicorn==0.16.1
psycopg2==2.4.5
dj-database-url==0.2.0
wsgiref==0.1.2
gevent==0.13.4
South==0.7.4
celery>=2.3.3
django-celery>=3.0.11

#Cache
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4
django-redis-cache==0.9.2
Markdown==2.0.3
django-picklefield

#Site Specific
django-registration == 0.7